+++ src/database/initialization/data_provider_to_json.sql
@@ -6,6 +6,7 @@ BEGIN
   RETURN jsonb_build_object(
     'shortCode', data_provider.short_code,
     'name', data_provider.name,
+    'keycloakOrganizationId', data_provider.keycloak_organization_id,
     'createdAt', data_provider.created_at
   );
 END;