+++ src/types/Funder.ts
@@ -1,11 +1,16 @@
 import { ajv } from '../ajv';
+import { KeycloakId, keycloakIdSchema } from './KeycloakId';
+import { Writable } from './Writable';
 import type { JSONSchemaType } from 'ajv';
-import type { Writable } from './Writable';
 import type { ShortCode } from './ShortCode';
 
 interface Funder {
 	readonly shortCode: ShortCode;
 	name: string;
+	// We do not really want "undefined" here, only null. See
+	// https://github.com/ajv-validator/ajv/issues/2283 and/or
+	// https://github.com/ajv-validator/ajv/issues/2163.
+	keycloakOrganizationId: KeycloakId | null | undefined;
 	readonly createdAt: string;
 }
 
@@ -18,6 +23,10 @@ const writableFunderSchema: JSONSchemaType<WritableFunder> = {
 	properties: {
 		name: {
 			type: 'string',
+		},
+		keycloakOrganizationId: {
+			...keycloakIdSchema,
+			nullable: true,
 		},
 	},
 	required: ['name'],